refac(timer): moved boss time display to its own component to avoid timer bloat
